Extracted GIExtGameplayInteractionContextWrapper in its own files
