+++ Source/GameplayInteractionsExtensions/Public/GIExtTypes.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <CoreMinimal.h>
-#include <GameplayInteractionContext.h>
+#include <SmartObjectRuntime.h>
 
 #include "GIExtTypes.generated.h"
 
@@ -33,47 +33,5 @@ struct GAMEPLAYINTERACTIONSEXTENSIONS_API FGIExtStartGameplayInteractionContext
 
     FSimpleDelegate PreActivationCallback;
     FSimpleDelegate OnInteractionFinishedCallback;
-    FOnSlotInvalidated  OnSlotInvalidatedCallback;
+    FOnSlotInvalidated OnSlotInvalidatedCallback;
 };
-
-UCLASS( BlueprintType )
-class GAMEPLAYINTERACTIONSEXTENSIONS_API UGIExtGameplayInteractionContextWrapper : public UObject, public FTickableGameObject
-{
-    GENERATED_BODY()
-
-public:
-    bool HasInteractionCompleted() const;
-    EGameplayInteractionAbortReason GetAbortReason() const;
-    bool StartInteraction( const FGIExtStartGameplayInteractionContext & context );
-    // bool TickInteraction( float delta_time );
-    void AbortInteraction( EGameplayInteractionAbortReason abort_reason );
-    void FinishInteraction();
-
-    void Tick( float delta_time ) override;
-    bool IsTickable() const override;
-    TStatId GetStatId() const override;
-
-    UFUNCTION( BlueprintCallable )
-    void SendEventToStateTree( const FGameplayTag tag );
-
-private:
-    USmartObjectSubsystem * GetSmartObjectSubsystem() const;
-    void OnSlotInvalidated( const FSmartObjectClaimHandle & claim_handle, ESmartObjectSlotState state );
-
-    UPROPERTY()
-    FGameplayInteractionContext GameplayInteractionContext;
-
-    FGIExtStartGameplayInteractionContext StartInteractionContext;
-    FGameplayInteractionAbortContext AbortContext;
-    bool bInteractionCompleted = false;
-};
-
-FORCEINLINE bool UGIExtGameplayInteractionContextWrapper::HasInteractionCompleted() const
-{
-    return bInteractionCompleted;
-}
-
-FORCEINLINE EGameplayInteractionAbortReason UGIExtGameplayInteractionContextWrapper::GetAbortReason() const
-{
-    return AbortContext.Reason;
-}
